fastapi_app/requirements.txt
# FastAPI 애플리케이션 요구사항

# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터 검증
pydantic==2.5.0
pydantic-settings==2.1.0

# PyTorch 및 Transformers
torch>=2.0.0
transformers>=4.35.0
kobert-tokenizer>=0.5.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 이미지 처리
Pillow>=10.0.0

# 유틸리티
python-dotenv==1.0.0
